requirements.txt
asgiref==3.7.2
Django==5.0.3
sqlparse==0.4.4
whitenoise==6.6.0  # must be in requirements.txt
gunicorn==21.2.0   # must be in requirements.txt
djangorestframework==3.14.0
